// examples/generated/org/example/Employee.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject } from 'emfts';
import type { Person } from './Person';

/**
 * Employee
 * @generated
 */
export interface Employee extends Person {
  employeeId: string;
  salary?: number;
  hireDate?: Date;
  supervisor?: Employee;

}